backend/core/requirements/base.txt
django==5.0.2
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-guardian>=2.4.0
django-tenants>=3.7.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.21.0
google-cloud-aiplatform>=1.42.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
